Changed some non-backwards compatible stuff.
Changed mails parameter to domains because it doesn't refer to actual emails but domains instead, makes more sense.
Added offsetWidth if you need to make adjustments to the tooltip width because the 'auto' logic doesn't fit it well in your design.
Updated demo using Bootstrap to make it more eyecandy.